move team to spaceship/base/client
de-duplicate team code in tunes and testflight client
