+++ spaceship/lib/spaceship/test_flight/client.rb
@@ -4,82 +4,6 @@ module TestFlight
       'https://itunesconnect.apple.com/testflight/v2/'
     end
 
-    # TODO: this is C&P from tunes_client
-    def user_details_data
-      return @_cached_user_details if @_cached_user_details
-      r = request(:get, '/WebObjects/iTunesConnect.woa/ra/user/detail')
-      @_cached_user_details = parse_response(r, 'data')
-    end
-
-
-    # TODO: this is C&P from tunes_client
-    # Fetches the User Detail information from ITC. This gets called often and almost never changes
-    # so we cache it
-    # @return [UserDetail] the response
-    def user_detail_data
-      @_cached_user_detail_data ||= Spaceship::Tunes::UserDetail.factory(user_details_data)
-    end
-
-
-
-    # TODO: this is C&P from tunes_client
-    # @return (Array) A list of all available teams
-    def teams
-      user_details_data['associatedAccounts'].sort_by do |team|
-        [
-          team['contentProvider']['name'],
-          team['contentProvider']['contentProviderId']
-        ]
-      end
-    end
-
-    # TODO: this is C&P from tunes_client
-    # @return (String) The currently selected Team ID
-    def team_id
-      return @current_team_id if @current_team_id
-
-      if teams.count > 1
-        puts "The current user is in #{teams.count} teams. Pass a team ID or call `select_team` to choose a team. Using the first one for now."
-      end
-      @current_team_id ||= teams[0]['contentProvider']['contentProviderId']
-    end
-
-    # TODO: this is C&P from tunes_client
-    # Set a new team ID which will be used from now on
-    def team_id=(team_id)
-      # First, we verify the team actually exists, because otherwise iTC would return the
-      # following confusing error message
-      #
-      #     invalid content provider id
-      #
-      available_teams = teams.collect do |team|
-        (team["contentProvider"] || {})["contentProviderId"]
-      end
-
-      result = available_teams.find do |available_team_id|
-        team_id.to_s == available_team_id.to_s
-      end
-
-      unless result
-        raise ITunesConnectError.new, "Could not set team ID to '#{team_id}', only found the following available teams: #{available_teams.join(', ')}"
-      end
-
-      response = request(:post) do |req|
-        req.url "ra/v1/session/webSession"
-        req.body = {
-          contentProviderId: team_id,
-          dsId: user_detail_data.ds_id # https://github.com/fastlane/fastlane/issues/6711
-        }.to_json
-        req.headers['Content-Type'] = 'application/json'
-      end
-
-      handle_itc_response(response.body)
-
-      @current_team_id = team_id
-    end
-
-
-
     # Returns an array of all available build trains (not the builds they include)
     def all_build_trains(app_id: nil, platform: nil)
       platform ||= "ios"
